fluids: Implement bc_slip
This is a weak version of slip boundary condition based on a Riemann
solve
